refactored: MainRecorder records Java data, includes vcml_to_fvsolve
