+++ scripts/local_build_native.sh
@@ -33,7 +33,9 @@ mvn clean install
 # run with native-image-agent to record configuration for native-image
 java -agentlib:native-image-agent=config-output-dir=target/recording \
      -jar target/vcell-native-1.0-SNAPSHOT.jar \
-     "$ROOT_DIR/vcell-native/src/test/resources/TinySpacialProject_Application0.xml" \
+     "$ROOT_DIR/vcell-native/src/test/resources/TinySpatialProject_Application0.xml" \
+     "$ROOT_DIR/vcell-native/src/test/resources/TinySpatialProject_Application0.vcml" \
+     "Simulation0" \
      "$ROOT_DIR/vcell-native/target/sbml-input"
 
 # build vcell-native as native shared object library
