+++ vcell-native/src/main/java/org/vcell/libvcell/Entrypoints.java
@@ -1,20 +1,5 @@
 package org.vcell.libvcell;
 
-import cbit.util.xml.VCLoggerException;
-import cbit.vcell.biomodel.BioModel;
-import cbit.vcell.geometry.GeometrySpec;
-import cbit.vcell.mapping.MappingException;
-import cbit.vcell.mapping.SimulationContext;
-import cbit.vcell.mongodb.VCMongoMessage;
-import cbit.vcell.parser.ExpressionException;
-import cbit.vcell.resource.PropertyLoader;
-import cbit.vcell.solver.Simulation;
-import cbit.vcell.solver.SolverException;
-import cbit.vcell.solver.TimeBounds;
-import cbit.vcell.solver.UniformOutputTimeSpec;
-import cbit.vcell.xml.XMLSource;
-import cbit.vcell.xml.XmlHelper;
-import cbit.vcell.xml.XmlParseException;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.graalvm.nativeimage.IsolateThread;
@@ -22,16 +7,12 @@ import org.graalvm.nativeimage.c.function.CEntryPoint;
 import org.graalvm.nativeimage.c.type.CCharPointer;
 import org.graalvm.nativeimage.c.type.CTypeConversion;
 import org.json.simple.JSONValue;
-import org.vcell.sbml.FiniteVolumeRunUtil;
-import org.vcell.sbml.vcell.SBMLExporter;
-import org.vcell.sbml.vcell.SBMLImporter;
 
-import java.beans.PropertyVetoException;
-import java.io.ByteArrayInputStream;
 import java.io.File;
-import java.io.FileInputStream;
-import java.io.InputStream;
 import java.util.concurrent.ConcurrentHashMap;
+
+import static org.vcell.libvcell.SolverUtils.sbmlToFiniteVolumeInput;
+import static org.vcell.libvcell.SolverUtils.vcmlToFiniteVolumeInput;
 
 
 public class Entrypoints {
@@ -123,70 +104,4 @@ public class Entrypoints {
         logger.info("Returning from sbmlToFiniteVolumeInput: " + json);
         return createString(json);
     }
-
-
-    public static void vcmlToFiniteVolumeInput(String vcml_content, String simulation_name, File outputDir) throws XmlParseException, MappingException, SolverException, ExpressionException {
-        GeometrySpec.avoidAWTImageCreation = true;
-        VCMongoMessage.enabled = false;
-        if (vcml_content.substring(0, 300).contains("<sbml xmlns=\"http://www.sbml.org/sbml")) {
-            throw new IllegalArgumentException("expecting VCML content, not SBML");
-        }
-        BioModel bioModel = XmlHelper.XMLToBioModel(new XMLSource(vcml_content));
-        bioModel.updateAll(false);
-        Simulation sim = bioModel.getSimulation(simulation_name);
-        if (sim == null) {
-            throw new IllegalArgumentException("Simulation not found: " + simulation_name);
-        }
-        FiniteVolumeRunUtil.writeInputFilesOnly(outputDir, sim);
-    }
-
-
-    public static void sbmlToFiniteVolumeInput(String sbml_content, File outputDir) throws MappingException, PropertyVetoException, SolverException, ExpressionException, VCLoggerException {
-        GeometrySpec.avoidAWTImageCreation = true;
-        VCMongoMessage.enabled = false;
-        SBMLExporter.MemoryVCLogger vcl = new SBMLExporter.MemoryVCLogger();
-        boolean bValidateSBML = true;
-        // input stream from sbml_content String
-        if (sbml_content.substring(0, 300).contains("<vcml xmlns=\"http://sourceforge.net/projects/vcell")) {
-            throw new IllegalArgumentException("expecting SBML content, not VCML");
-        }
-        InputStream is = new ByteArrayInputStream(sbml_content.getBytes());
-        SBMLImporter importer = new SBMLImporter(is, vcl, bValidateSBML);
-        BioModel bioModel = importer.getBioModel();
-        bioModel.updateAll(false);
-
-        final double duration = 5.0;  // endpoint arg
-        final double time_step = 0.1;  // endpoint arg
-        //final ISize meshSize = new ISize(10, 10, 10);  // future endpoint arg
-        SimulationContext simContext = bioModel.getSimulationContext(0);
-        Simulation sim = new Simulation(simContext.getMathDescription(), simContext);
-        sim.getSolverTaskDescription().setTimeBounds(new TimeBounds(0.0, duration));
-        sim.getSolverTaskDescription().setOutputTimeSpec(new UniformOutputTimeSpec(time_step));
-
-        FiniteVolumeRunUtil.writeInputFilesOnly(outputDir, sim);
-    }
-
-
-
-    // Input Goes as Follows: SBML Input, Output dir
-    //    "/Users/evalencia/Documents/VCell_Repositories/vcell/vcell-rest/src/test/resources/TinySpacialProject_Application0.xml"
-    //    "/Users/evalencia/Documents/VCell_Repositories/vcell/vcell-nativelib/target/sbml-input"
-    public static void main(String[] args) {
-        try {
-            logger.info("Logger logging");
-            PropertyLoader.setProperty(PropertyLoader.vcellServerIDProperty, "none");
-            PropertyLoader.setProperty(PropertyLoader.mongodbDatabase, "none");
-            File sbml_file = new File(args[0]);
-            // read sbml_file and create a string object
-            try (FileInputStream fis = new FileInputStream(sbml_file)) {
-                byte[] data = fis.readAllBytes();
-                logger.info("Read " + data.length + " bytes from " + sbml_file);
-                String sbml_str = new String(data);
-                sbmlToFiniteVolumeInput(sbml_str, new File(args[1]));
-            }
-        } catch (Exception e) {
-            System.out.println(e.getMessage());
-            logger.error("Error processing spatial model", e);
-        }
-    }
-}
+ }
