+++ vcell-native/src/test/java/org/vcell/libvcell/EntrypointsTest.java
@@ -16,6 +16,8 @@ import java.util.stream.Collectors;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.vcell.libvcell.SolverUtils.sbmlToFiniteVolumeInput;
+import static org.vcell.libvcell.SolverUtils.vcmlToFiniteVolumeInput;
 
 public class EntrypointsTest {
 
@@ -23,7 +25,7 @@ public class EntrypointsTest {
     public void testSbmlToFiniteVolumeInput() throws PropertyVetoException, SolverException, ExpressionException, MappingException, VCLoggerException, IOException {
         String sbmlContent = getFileContentsAsString("/TinySpatialProject_Application0.xml");
         Path output_dir = Files.createTempDirectory("sbmlToFiniteVolumeInput");
-        Entrypoints.sbmlToFiniteVolumeInput(sbmlContent, output_dir.toFile());
+        sbmlToFiniteVolumeInput(sbmlContent, output_dir.toFile());
     }
 
     // TODO: better exception handling by JSBML needed
@@ -31,7 +33,7 @@ public class EntrypointsTest {
     public void testSbmlToFiniteVolumeInput_not_well_formed() throws IOException {
         String sbmlContent = getFileContentsAsString("/TinySpatialProject_Application0.xml").replaceAll("<model ", "<modelXYZ ");
         Path output_dir = Files.createTempDirectory("sbmlToFiniteVolumeInput");
-        ClassCastException exc = assertThrows(ClassCastException.class, () -> Entrypoints.sbmlToFiniteVolumeInput(sbmlContent, output_dir.toFile()));
+        ClassCastException exc = assertThrows(ClassCastException.class, () -> sbmlToFiniteVolumeInput(sbmlContent, output_dir.toFile()));
         assertEquals(
                 "class org.sbml.jsbml.xml.XMLNode cannot be cast to class org.sbml.jsbml.Annotation " +
                 "(org.sbml.jsbml.xml.XMLNode and org.sbml.jsbml.Annotation are in unnamed module of loader 'app')",
@@ -42,7 +44,7 @@ public class EntrypointsTest {
     public void testSbmlToFiniteVolumeInput_vcml_instead() throws IOException {
         String sbmlContent = getFileContentsAsString("/TinySpatialProject_Application0.vcml");
         Path output_dir = Files.createTempDirectory("sbmlToFiniteVolumeInput");
-        IllegalArgumentException exc = assertThrows(IllegalArgumentException.class, () -> Entrypoints.sbmlToFiniteVolumeInput(sbmlContent, output_dir.toFile()));
+        IllegalArgumentException exc = assertThrows(IllegalArgumentException.class, () -> sbmlToFiniteVolumeInput(sbmlContent, output_dir.toFile()));
         assertEquals("expecting SBML content, not VCML", exc.getMessage());
     }
 
@@ -51,7 +53,7 @@ public class EntrypointsTest {
         String vcmlContent = getFileContentsAsString("/TinySpatialProject_Application0.vcml");
         Path output_dir = Files.createTempDirectory("vcmlToFiniteVolumeInput");
         String simulationName = "Simulation0";
-        Entrypoints.vcmlToFiniteVolumeInput(vcmlContent, simulationName, output_dir.toFile());
+        vcmlToFiniteVolumeInput(vcmlContent, simulationName, output_dir.toFile());
     }
 
     @Test
@@ -60,7 +62,7 @@ public class EntrypointsTest {
         Path output_dir = Files.createTempDirectory("vcmlToFiniteVolumeInput");
         String simulationName = "wrong_sim_name";
         // expect to throw an IllegalArgumentException
-        IllegalArgumentException exc = assertThrows(IllegalArgumentException.class, () -> Entrypoints.vcmlToFiniteVolumeInput(vcmlContent, simulationName, output_dir.toFile()));
+        IllegalArgumentException exc = assertThrows(IllegalArgumentException.class, () -> vcmlToFiniteVolumeInput(vcmlContent, simulationName, output_dir.toFile()));
         assertEquals("Simulation not found: wrong_sim_name", exc.getMessage());
     }
 
@@ -71,7 +73,7 @@ public class EntrypointsTest {
         Path output_dir = Files.createTempDirectory("vcmlToFiniteVolumeInput");
         String simulationName = "wrong_sim_name";
         // expect to throw an IllegalArgumentException
-        RuntimeException exc = assertThrows(RuntimeException.class, () -> Entrypoints.vcmlToFiniteVolumeInput(vcmlContent, simulationName, output_dir.toFile()));
+        RuntimeException exc = assertThrows(RuntimeException.class, () -> vcmlToFiniteVolumeInput(vcmlContent, simulationName, output_dir.toFile()));
         assertEquals(
                 "source document is not well-formed\n" +
                         "Error on line 25: The element type \"SimpleReactionXYZ\" must be terminated by the matching end-tag \"</SimpleReactionXYZ>\".",
@@ -84,7 +86,7 @@ public class EntrypointsTest {
         Path output_dir = Files.createTempDirectory("vcmlToFiniteVolumeInput");
         String simulationName = "wrong_sim_name";
         // expect to throw an IllegalArgumentException
-        IllegalArgumentException exc = assertThrows(IllegalArgumentException.class, () -> Entrypoints.vcmlToFiniteVolumeInput(vcmlContent_actually_sbml, simulationName, output_dir.toFile()));
+        IllegalArgumentException exc = assertThrows(IllegalArgumentException.class, () -> vcmlToFiniteVolumeInput(vcmlContent_actually_sbml, simulationName, output_dir.toFile()));
         assertEquals("expecting VCML content, not SBML", exc.getMessage());
     }
